finalized warehouse service to accomodate producing and consuming events
for more details, the script no longer forces the container to close if the consumer topic being listened to, do not exists in Kafka. Instead, it showcases a warning log and tries again after 5 min. Other than that, the approach for the warehouse is as such. 1) A number of buckets is appended to a list depending on how much capacity is allowed in the warehouse. This is the state. 2) An event is initially produced at script execution and sent to the warehouse topic, that notifies a bucket is available in the warehouse for pickup. 3) While it does that, it listens to the mqtt-mediator topic, to consume events that determines if a full production cycle has been finished. 3) If a cycle has been finished, two scenarios can happen. First, if there is no buckets to pick up in the warehouse, it states the scenario as a log, and closes the subscription for consuming cycles. Second, if there are still buckets left, it produces yet another notification, logs the amount left, and removes a bucket from the initial list that contained the state of buckets in the warehouse.